Download DisMAP Raster

Download a single interpolated biomass raster by slice ID

Usage

download_dismap_raster(
  img_url,
  slice_id,
  out_tif = NULL,
  transform = function(x) x^(1/3),
  out_sr = NULL,
  bbox = NULL,
  bbox_sr = NULL,
  overwrite = F,
  verbose = F
)

Arguments

img_url

The URL for the image server for a given region

slice_id

The slice ID to download

out_tif

optional output GeoTIFF path to write raster

transform

Function to transform raster values (default is cube root): function(x) x^(1/3)

out_sr

optional output spatial reference for output raster; defaults to native spatial reference

bbox

optional bounding box coordinates (xmin, ymin, xmax, ymax); defaults to maximum extent

bbox_sr

optional spatial reference for bounding box; defaults to native spatial reference; you can use geographic (4326)

overwrite

if out_tif exists, overwrite it (default is FALSE)

verbose

if TRUE, print additional information

Value

A terra::rast object