Get DisMAP layer names, given region season

Get a list of available layers for a given region and season

Usage

get_dm_dataset_layers(dataset_code)

Arguments

datset_code

The region (and season) code, per available dataset_code in dm_regions

Value

A character vector of layer names, usually species scientific name or "Species Richness"

Examples

get_dm_dataset_layers("AI")
#>  [1] "Anoplopoma fimbria"                     
#>  [2] "Aphrocallistes vastus"                  
#>  [3] "Atheresthes stomias and A evermanni"    
#>  [4] "Axinella blanca"                        
#>  [5] "Bathymaster signatus"                   
#>  [6] "Bathyraja spp"                          
#>  [7] "Berryteuthis magister"                  
#>  [8] "Callogorgia compressa"                  
#>  [9] "Cheiraster Luidiaster dawsoni"          
#> [10] "Chionoecetes bairdi"                    
#> [11] "Core Species Richness"                  
#> [12] "Crossaster papposus"                    
#> [13] "Cucumaria fallax"                       
#> [14] "Diplopteraster multipes"                
#> [15] "Elassochirus cavimanus"                 
#> [16] "Enteroctopus dofleini"                  
#> [17] "Eumicrotremus orbis"                    
#> [18] "Fusitriton oregonensis"                 
#> [19] "Gadus chalcogrammus"                    
#> [20] "Gadus macrocephalus"                    
#> [21] "Glyptocephalus zachirus"                
#> [22] "Gorgonocephalus eucnemis"               
#> [23] "Gymnocanthus galeatus"                  
#> [24] "Halocynthia aurantium"                  
#> [25] "Hemilepidotus jordani"                  
#> [26] "Hippoglossoides elassodon"              
#> [27] "Hippoglossus stenolepis"                
#> [28] "Histodermella kagigunensis"             
#> [29] "Hyas lyratus"                           
#> [30] "Lepidopsetta sp"                        
#> [31] "Lethasterias nanimensis"                
#> [32] "Lithodes aequispinus"                   
#> [33] "Malacocottus zonurus"                   
#> [34] "Microstomus pacificus"                  
#> [35] "Monanchora pulchra"                     
#> [36] "Mycale Mycale loveni"                   
#> [37] "Myoxocephalus polyacanthocephalus"      
#> [38] "Ophiopholis aculeata"                   
#> [39] "Oregonia gracilis"                      
#> [40] "Pagurus brandti"                        
#> [41] "Pandalus eous"                          
#> [42] "Pandalus tridens"                       
#> [43] "Pleurogrammus monopterygius"            
#> [44] "Pododesmus macrochisma"                 
#> [45] "Podothecus accipenserinus"              
#> [46] "Pteraster militaris"                    
#> [47] "Rossia pacifica"                        
#> [48] "Sarritor frenatus"                      
#> [49] "Sebastes alutus"                        
#> [50] "Sebastes borealis"                      
#> [51] "Sebastes melanostictus and S aleutianus"
#> [52] "Sebastes polyspinis"                    
#> [53] "Sebastes variabilis and S ciliatus"     
#> [54] "Sebastolobus alascanus"                 
#> [55] "Species Richness"                       
#> [56] "Tedania Tedania kagalaskai"             
#> [57] "Triglops forficatus"                    
#> [58] "Triglops scepticus"                     
#> [59] "Zaprora silenus"