Get DisMAP raster for dataset layer year

Download a single interpolated biomass raster by slice ID

Usage

get_dm_raster(
  dataset_code,
  layer,
  year,
  out_tif = NULL,
  out_sr = NULL,
  transform = function(x) x^(1/3),
  bbox = NULL,
  bbox_sr = NULL,
  overwrite = F,
  verbose = F
)

Arguments

dataset_code

The DisMAP dataset code (e.g., "AI", "EBS", "GOA"), per dm_datasets$dataset_code

layer

The layer in the dataset, typically the scientific name of the species or "Species Richness", per get_dm_dataset_layers(dataset_code)

year

The year of the slice to download, per get_dm_dataset_layer_years(dataset_code, layer)

out_tif

optional output GeoTIFF path to write raster

out_sr

optional output spatial reference for output raster; defaults to native spatial reference

transform

Function to transform raster values (default is cube root): function(x) x^(1/3)

bbox

optional bounding box coordinates (xmin, ymin, xmax, ymax); defaults to maximum extent

bbox_sr

optional spatial reference for bounding box; defaults to native spatial reference; you can use geographic (4326)

overwrite

if out_tif exists, overwrite it (default is FALSE)

verbose

if TRUE, print additional information

Value

A terra::rast object

Examples

get_dm_raster("AI", "Species Richness", 2015) |> terra::plet()
#> Loading required package: httr2
#> Loading required package: terra
#> terra 1.8.42
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plet': year must be one of the dataset's available years: 1991, 1994, 1997, 2000, 2002, 2004, 2006, 2010, 2012, 2014, 2016, 2018, 2022
get_dm_raster("NEUS_SPR", "Paralichthys dentatus", 2015, out_tif = "ne_spring_summer_flounder_2015.tif")  |> terra::plet()