Plot DisMAP Distribution Map

Create a static or animated plot of species distribution

Usage

plot_dismap_distribution(
  data,
  value_col = "transformed_wtcpue",
  animated = FALSE,
  xlim = NULL,
  ylim = NULL,
  color_palette = NULL
)

Arguments

data

Data frame with lon, lat, year, and transformed biomass columns

value_col

Name of column with values to plot

animated

Whether to create an animated plot

xlim, ylim

Limits for x and y axes

color_palette

Color palette function

Value

A ggplot object or animated gif